# swaggerSpec.yml
openapi: 3.0.0
info:
  title: Cat Management API
  description: This is an API to manage a collection of cats, including adding, updating, and retrieving cat details and types.
  version: 1.0.0
  contact:
    name: Uranbek Anarbaev
    email: uranbek@example.com

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: Cats
    description: Operations related to managing cats
  - name: Types
    description: Operations related to managing cat types

paths:
  /gettypes:
    get:
      tags:
        - Types
      summary: Get all cat types
      description: Retrieve a list of all available cat types.
      responses:
        '200':
          description: A list of cat types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypeRead'
  
  /gettypes/{type}:
    get:
      tags:
        - Types
      summary: Get cats by type
      description: Retrieve a list of cats filtered by their type.
      parameters:
        - in: path
          name: type
          schema:
            type: string
          required: true
          description: The type of the cat to retrieve.
      responses:
        '200':
          description: A list of cats of the specified type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KittyRead'

  /cat:
    get:
      tags:
        - Cats
      summary: Get all cats
      description: Retrieve a list of all cats, optionally filtered by type.
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: Filter cats by their type.
      responses:
        '200':
          description: A list of cats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KittyRead'
    
    post:
      tags:
        - Cats
      summary: Create a new cat
      description: Add a new cat to the collection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KittyCreate'
      responses:
        '201':
          description: The created cat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KittyRead'

  /cat/{cat_id}:
    get:
      tags:
        - Cats
      summary: Get a specific cat
      description: Retrieve details of a specific cat by its ID.
      parameters:
        - in: path
          name: cat_id
          schema:
            type: integer
          required: true
          description: The ID of the cat to retrieve.
      responses:
        '200':
          description: Details of the specified cat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KittyRead'
        '404':
          description: Cat not found
    
    put:
      tags:
        - Cats
      summary: Update a specific cat
      description: Update details of an existing cat by ID.
      parameters:
        - in: path
          name: cat_id
          schema:
            type: integer
          required: true
          description: The ID of the cat to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KittyCreate'
      responses:
        '200':
          description: The updated cat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KittyRead'
        '404':
          description: Cat not found

    delete:
      tags:
        - Cats
      summary: Delete a specific cat
      description: Delete a specific cat by its ID.
      parameters:
        - in: path
          name: cat_id
          schema:
            type: integer
          required: true
          description: The ID of the cat to delete.
      responses:
        '200':
          description: The deleted cat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KittyRead'
        '404':
          description: Cat not found

components:
  schemas:
    KittyCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        age:
          type: integer
        color:
          type: string
        type:
          type: string
      required:
        - name
        - description
        - age
        - color
        - type

    KittyRead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        age:
          type: integer
        color:
          type: string
        type:
          type: string
      required:
        - id
        - name
        - description
        - age
        - color
        - type

    TypeRead:
      type: object
      properties:
        type:
          type: string
      required:
        - type